Refactor: extract shared Docs API types and URL helper
- Move docsBatchUpdateRequest and related types from update_document.go
  to docs_types.go, since they're shared with create_document.go
- Extract documentEditURL() helper to eliminate duplicated URL
  construction in create_document.go and get_document.go
- All tests pass including race detector
